[bug-69563] try to avoid int overflow in Zip64 code (probably just moves the problem somewhere else though)
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@1923647 13f79535-47bb-0310-9956-ffa450edef68

+++ poi-ooxml/src/main/java/org/apache/poi/xssf/streaming/Zip64Impl.java
@@ -46,8 +46,8 @@ class Zip64Impl {
         final String filename;
         long crc;
         long size;
-        int compressedSize;
-        int offset;
+        long compressedSize;
+        long offset;
 
         Entry(String filename) {
             this.filename = filename;
@@ -128,7 +128,7 @@ class Zip64Impl {
     /**
      * Write End of central directory record (EOCD)
      */
-    int writeEND(int entriesCount, int offset, int length) throws IOException {
+    int writeEND(int entriesCount, long offset, long length) throws IOException {
         written = 0;
         writeInt(PK0506);         // "PK\005\006"
         writeShort(0);         // number of this disk
